bin/requirements.txt
-i https://pypi.org/simple
gitdb2==2.0.5
gitpython==2.1.11
pyyaml==5.1
smmap2==2.0.5
